created pseudo-module for pydicom3 and import from there
